% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adj_asd.R
\name{adj_asd}
\alias{adj_asd}
\title{adj_asd}
\usage{
adj_asd(Xi, iadj)
}
\description{
adj_asd
}
\section{Elimine les sauts dus au passage d'un dtecteur  a un autre.}{
On extrapole l'ajustement linaire fait sur les \link{ws} points avant le saut, sur le point suivant
La diffrence entre l'ancien et le nouveau "point suivant" est soustraite a tous les points apres le "point suivant"
\link{651,1451}
}

