% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adj_NaturaSpec.R
\name{adj_NaturaSpec}
\alias{adj_NaturaSpec}
\title{adj_NaturaSpec}
\usage{
adj_NaturaSpec(X)
}
\description{
adj_NaturaSpec
}
\section{Elimine les sauts dus au passage d'un dtecteur  a un autre.}{
On extrapole l'ajustement linaire fait sur les \link{ws} points avant le saut, sur les 14 points suivants
(interpolated after gap, I guess from SpectralEvolution)
La difference entre le point à gap + 15 (entre vrai spectre et fit) est soustraite a tous les points suivants
\link{632,?}
}

