% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridcv_mec.R
\name{.mat}
\alias{.mat}
\title{gridcv_mec}
\usage{
.mat(X, prefix = NULL)
}
\description{
Reprende gridcv, mais utilise les foncions gcv et gcvlv au lieu de gridcv et gridcvlv de rchemo
Ce qui permet de rajouter les prédictions pour chaque indiv
}
