% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridscore_mec.R
\name{gscore_mec}
\alias{gscore_mec}
\title{gridscore_mec}
\usage{
gscore_mec(Xtrain, Ytrain, X, Y, score, fun, pars, verb = FALSE)
}
\description{
Reprende gridscore, mais utilise les foncions gscore_mec et gscorelv_mec au lieu de gscore et gscorelv
}
