% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridscore_mec.R
\name{gscorelv_mec}
\alias{gscorelv_mec}
\title{gscorelv_mec}
\usage{
gscorelv_mec(Xtrain, Ytrain, X, Y, score, fun, nlv, pars = NULL, verb = FALSE)
}
\description{
Reprende gridscore, mais utilise les foncions gscore_mec et gscorelv_mec au lieu de gscore et gscorelv
}
