% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_fac_rnirs.R
\name{table_fac_rnirs}
\alias{table_fac_rnirs}
\title{Table_fac_rnirs}
\usage{
table_fac_rnirs(fm, dat_fac, numfac, segm, comp, rep = 1:max(fm$y$rep))
}
\arguments{
\item{fac}{}
}
\description{
Table_fac_rnirs
}
